+++ lib/appsignal/instrumentation/helpers.ex
@@ -8,6 +8,15 @@ defmodule Appsignal.Instrumentation.Helpers do
   @type instrument_arg :: Transaction.t | Plug.Conn.t | pid()
 
   @doc """
+  Execute the given function in start / finish event calls in the current
+  transaction. See `instrument/6`.
+  """
+  @spec instrument(String.t, String.t, function) :: any
+  def instrument(name, title, function) do
+    instrument(self(), name, title, "", function)
+  end
+
+  @doc """
   Execute the given function in start / finish event calls. See `instrument/6`.
   """
   @spec instrument(instrument_arg, String.t, String.t, function) :: any
@@ -34,7 +43,7 @@ defmodule Appsignal.Instrumentation.Helpers do
   import Appsignal.Instrumentation.Helpers, only: [instrument: 4]
 
   def index(conn, _params) do
-    result = instrument conn, "net.http", "Some slow backend call", fn() ->
+    result = instrument "net.http", "Some slow backend call", fn() ->
       Backend.get_result()
     end
     json conn, result
@@ -52,17 +61,10 @@ defmodule Appsignal.Instrumentation.Helpers do
     end
   end
 
-  if Appsignal.phoenix? do
-    def instrument(%Plug.Conn{} = conn, name, title, body, body_format, function) do
-      instrument(conn.assigns.appsignal_transaction, name, title, body, body_format, function)
-    end
-  end
-
   def instrument(%Transaction{} = transaction, name, title, body, body_format, function) do
     Transaction.start_event(transaction)
     result = function.()
     Transaction.finish_event(transaction, name, title, body, body_format)
     result
   end
-
 end